+++ tailwind.config.js
@@ -1,5 +1,3 @@
-process.env.TAILWIND_MODE = isProductionMode() ? 'build' : 'watch';
-
 module.exports = {
   mode: 'jit',
   purge: ['./src/**/*.{html,ts,scss}', './projects/**/*.{html,ts,scss}'],
@@ -12,9 +10,3 @@ module.exports = {
   },
   plugins: [require('@tailwindcss/forms')],
 };
-
-function isProductionMode() {
-  const argv = process.argv.join(' ').toLowerCase();
-  const isProdEnv = process.env.NODE_ENV === 'production';
-  return isProdEnv || [' build', ':build', ' deploy', 'ng b', '--prod'].some((command) => argv.includes(command));
-}